% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_sample_vf.R
\docType{data}
\name{ga_sample_vf}
\alias{ga_sample_vf}
\title{GA Voterfile sample}
\format{
A data frame with 1000 rows and 10 variables:
\describe{
\item{\code{county_code}}{integer, internal GA county code}
\item{\code{surname}}{character lastname/surname of the voter}
\item{\code{race}}{character race code for voter}
\item{\code{race_desc}}{character race for voter detailed}
\item{\code{gender}}{character voter's gender}
\item{\code{block}}{character census block 4 digit ID}
\item{\code{tract}}{character census tract 6 digit ID}
\item{\code{county}}{character census county fips code ID}
\item{\code{state}}{character state postal abbreviation}
\item{\code{residence_zipcode}}{character the five digit zip code for the voter}
}
}
\source{
Georgia Election Commission 2020
}
\usage{
ga_sample_vf
}
\description{
A sample of observations from the Georiga voterfile with block level identifiers
}
\keyword{datasets}
