% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_bisg_geo.R
\name{merge_bisg_geo}
\alias{merge_bisg_geo}
\title{Merge geographic tabular data command}
\usage{
merge_bisg_geo(df, geo_df, vec_merge)
}
\arguments{
\item{df}{The data frame for the voter file.}

\item{geo_df}{The demographic data with census/other geographic identifiers, and r_eth fields that all sum to 1 by column
(i.e. the proportion of a given race that lives in the geographic unit relative to the rest of the state.)}

\item{vec_merge}{The vector with the character fields necessary to merge. Should be identical in both}
}
\value{
A dataframe with the merged voterfile and geographic demographic data necessary to run BISG via predict_race_any
}
\description{
This function merges on user provided demographic data with geographic codes, and is incorporated into the predict_race_any cmd
}
\examples{
ga_data <-  zipWRUext2::ga_sample_vf
ga_block_sample <- zipWRUext2::ga_sample_blocks
ga_data2 <- zip_wru(ga_data, ga_block_sample,  c("county","tract","block"))

}
