% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_surnames.R
\name{merge_surnames}
\alias{merge_surnames}
\title{Command to merge surname dictionary from wru onto data}
\usage{
merge_surnames(
  voter.file,
  surname.year = 2010,
  clean.surname = T,
  impute.missing = T
)
}
\arguments{
\item{voter.file}{The data frame for the voter file.}

\item{surname.year}{The year for the surname dictionary; takes form of 2000 or 2010; defaults to 2010}

\item{clean.surname}{Triggers script to clean the surname field}

\item{impute.missing}{Tells script to impute missing surnames}
}
\value{
A dataframe with pred.eth values of the predicted probability for a given race. These rowsum to 1.
}
}
\description{
This function is grandfathered in from a previous version of	Kabir Khanna and Kosuke Imai's wru package from 2021; necessary to
operate the predict_race_any and zip_wru commands.
}
\examples{
surname <- c("RUSSELL", "JONES", "SMITH", "ALLEN")
y <- c("31302", "30339", "30040", "30030")
df <- data.frame(surname,y)
merge_surnames(df)
  surname     y surname.match  p_whi  p_bla  p_his  p_asi  p_oth
3 RUSSELL 31302       RUSSELL 0.7705 0.1669 0.0246 0.0057 0.0323
2   JONES 30339         JONES 0.5519 0.3848 0.0229 0.0044 0.0361
4   SMITH 30040         SMITH 0.7090 0.2311 0.0240 0.0050 0.0308
1   ALLEN 30030         ALLEN 0.6759 0.2617 0.0247 0.0054 0.0324


}
