% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_race_any.R
\name{predict_race_any}
\alias{predict_race_any}
\title{Predict race via BISG with user provided data}
\usage{
predict_race_any(df, geo_df, vec_merge)
}
\arguments{
\item{df}{The data frame for the voter file.}

\item{geo_df}{The demographic data with census/other geographic identifiers, and r_eth fields that all sum to 1 by column
(i.e. the proportion of a given race that lives in the geographic unit relative to the rest of the state.)}

\item{vec_merge}{The vector with the character fields necessary to merge. Should be identical in both}
}
\value{
A dataframe with pred.eth values of the predicted probability for a given race. These rowsum to 1.
}

\description{
This function predicts race via BISG with user provided data, and makes use of the merge_bisg_geo command to do so.
}
\examples{
ga_data <-  zipWRUext2::ga_sample_vf
ga_block_sample <- zipWRUext2::ga_sample_blocks
ga_data2 <- predict_race_any(ga_data, ga_block_sample,  c("county","tract","block"))

}
