% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_herfindahl.R
\name{race_herfindahl_scores}
\alias{race_herfindahl_scores}
\title{Race Confidence Scorer}
\usage{
race_herfindahl_scores(
  dataframe1,
  predNames = c("pred.whi", "pred.bla", "pred.his", "pred.asi", "pred.oth")
)
}
\arguments{
\item{dataframe1}{The data frame with the predicted race fields}

\item{predNames}{The vector of five estimated race fields from the zip_wru cmd. Defaults to pred.eth, e.g. pred.whi.}
}
\value{
A dataframe with the herfindahl scores to measure certainty of individual estimate, highest race estimate, and plurality race.
}
\description{
This function takes the predicted race scores, finds the plurality race, then finds the degree of confidence that the user can have in race via
the herfindahl index.
}
\examples{
wi_data <- zipWRUext2::wi_data
wi_data2 <- zip_wru(wi_data, state="WISCONSIN", type1="census", year1="2010", zip_col="zcta5", surname_field = "lastname")
wi_data2 <- race_herfindahl_scores(wi_data2)
head(wi_data4[,26:28])
  herf_weight max_race_prob plural_race
1   0.9898037     0.9948803       white
2   0.9407473     0.9696928       white
3   0.9858093     0.9928664       white
4   0.9471131     0.9730675       white
5   0.8995676     0.9480736       white
6   0.9725417     0.9861357       white

}
