% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_preserve_sum.R
\name{round_preserve_sum}
\alias{round_preserve_sum}
\title{Round data while preserving the total}
\usage{
round_preserve_sum(x, digits = 0)
}
\arguments{
\item{x}{The vector of the number to be rounded.}

\item{digits}{The number of digits to round to.}
}
\value{
A vector that rounds numbers while preserving the total.
}
}
\description{
This function rounds numbers in a vector to the specified num of digits, while ensuring it remains equal to total.
}
\examples{
ga_data <-  zipWRUext2::ga_sample_vf
ga_block_sample <- zipWRUext2::ga_sample_blocks
ga_data2 <- predict_race_any(ga_data, ga_block_sample,  c("county","tract","block"))

}
