% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wru_aggregate_pres.R
\name{wru_aggregate_pres}
\alias{wru_aggregate_pres}
\title{Aggregate weighted racial estimates and preserve total}
\usage{
wru_aggregate_pres(
  df,
  group_var = "zcta5",
  predNames = c("pred.whi", "pred.bla", "pred.his", "pred.asi", "pred.oth")
)
}
\arguments{
\item{df}{The data frame for the voter file.}

\item{predNames}{The predicted prob columns that the bisg predict commands estimated. Defaults to pred.eth, e.g. pred.whi}

\item{geo_df}{The grouping variable to collapse by; have in quotes to be read in as a string.}
}
\value{
A dataframe gropued by the user specified variable with whole numbers for a given racial category
}
}
\description{
This function aggregates by the user specified grouping varaible for the weighted probabilities of the racial categories, while also
ensuring that the rounded estimates adhere to the total number.
}
\examples{
wi_data <- zipWRUext2::wi_data
wi_data4 <- zip_wru(wi_data, state="WISCONSIN",  year1="2010", zip_col="zcta5", surname_field = "lastname")
wi_data4col <- wru_aggregate_pres(wi_data4, group_var = "zcta5")
head(wi_data4col)
A tibble: 6 x 7
zcta5 pred.asi pred.bla pred.his pred.oth pred.whi total_pop
<chr>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>     <dbl>
 1 53001        0        0        0        0        1         1
2 53004        0        0        0        0        1         1
3 53005        0        0        0        0        3         3
4 53006        0        0        0        0        1         1
5 53007        0        0        0        0        1         1
6 53011        0        0        0        0        1         1

}
