% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wru_cis.R
\name{wru_cis}
\alias{wru_cis}
\title{BISG confidence intervals}
\usage{
wru_cis(
  df,
  group_var = "zcta5",
  predNames = c("pred.whi", "pred.bla", "pred.his", "pred.asi", "pred.oth"),
  ci_vec = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{df}{The data frame outputted by zip_wru or a related BISG command}

\item{group_var}{The geographic grouping variable that will be aggregated to; string input and defaults to zcta5}

\item{predNames}{The vector of pred. names that reflect the BISG race estimates. Defaults to "pred.whi", pred.bla, etc.}

\item{ci_vec}{The vector of probabilities that will be used to create the confidence intervals. Defaults to the 95\% CI range}
}
\value{
A dataframe with the pred.race + confidence intervals \%s in the aggregated format.
}

\description{
This function takes the BISG dataframe output and computes confidence intervals for a given grouped level of geography
}
\examples{
wi_data <-  zipWRUext2::wi_data
wi_data2 <- zip_wru(wi_data, state="WISCONSIN",  year1="2010", zip_col="zcta5", surname_field = "lastname")
wi_data2_cis <- wru_cis(wi_data2, group_var = "zcta5")

}
