% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_wru.R
\name{zip_wru}
\alias{zip_wru}
\title{BISG with ZIP codes}
\usage{
zip_wru(
  dataframe1,
  state,
  type1 = "census",
  year1 = "2010",
  zip_col = "zcta5",
  surname_field = "surname"
)
}
\arguments{
\item{dataframe1}{The data frame for the voter file.}

\item{state}{The state name for the voter file, in all upper case letters}

\item{type1}{The type of the data to use for BISG. Can be either census or acs}

\item{year1}{The year for the data to use for the demographic population data. If 2010, then type1 must be equal to census. Spans 2010 - 2018}

\item{zip_col}{The column name for the five digit zip code field within the dataframe. Defaults to zcta5}

\item{surname_field}{The column name for the surname that the wru package will use to merge on the surname probabilities. Note that they
must all be uppercase, and if not already, the column will be renamed to surname}
}
\value{
A dataframe with the pred.race fields for the probability that an individual is of a given race. These sum to 1.
}
}
\description{
This function takes voter file data set up for wru and performs BISG at teh ZIP code level, incorporating Census 2010 or ACS 2011 - 2018
data, as acquired from Social Explorer.
}
\examples{
wi_data <- wi_data <- zipWRUext::wi_data
wi_data2 <- zip_bisg(wi_data, state="WISCONSIN", type1="census", year1="2010", zip_col="zcta5", surname_field = "lastname")

}
