% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wivoterfile.R
\docType{data}
\name{wi_data}
\alias{wi_data}
\title{WI Voterfile sample}
\format{
A data frame with 900 rows and 3 variables:
\describe{
\item{\code{lastname}}{character lastname/surname of the voter}
\item{\code{county}}{character county of voter}
\item{\code{zcta5}}{character the five digit zip code for the voter
}
}
\source{
Wisconsin Election Commission 2020
}
\usage{
wi_data
}
\description{
A sample of observations from the Wisconsin voterfile in order to run BISG
}
\keyword{datasets}
